backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
google-cloud-vision
google-cloud-language
python-dotenv
pydantic
python-multipart
pydantic-settings
google-genai
httpx
watchfiles
google-api-python-client
httpx
